Extract the mint method to hooks
