+++ src/constants/poster.ts
@@ -1,1 +1,2 @@
 export const MINT_DATE = '2022-06-18T23:00:00-04:00'; // time is in EST (GMT-04:00)
+export const NFT_TOKEN_ID = 0;
